Move the area effect cloud and lingering potion entities to the new package
